refactor: break component youtube video
